+++ packages/scripts/src/utils/contracts.ts
